Fix RECURRENCE-ID appearing before RRULE
Some ical files have a specific RECURRENCE-ID override entry that
appears before the more general RRULE parent entry for the same UID.
There were two bugs related to this:
1) If multiple entries for the same UID appear, the later UID entries
were supposed to overwrite fields from the previous UID entries.  This
wasn't working correctly, so fields weren't getting overwritten.  In
this case, it meant that the RECURRENCE-ID parent entry wasn't getting
the RRULE added to it.
2) The correct behaviour (as far as I can tell) should be for the parent
record to look the same, regardless of which entry comes first, and the
RECURRENCE-ID entry should end up in the recurrences array.  This is now
handled by letting the RECURRENCE-ID entry be the parent *and* an entry
in the array, but then letting the RRULE example overwrite the fields in
the parent as appropriate.

Added a test to demonstrate the case and the correct processing of it.