Fix comma-separated EXDATES, change ISO string keys
Added support for comma-separated EXDATES, since those are allowed by the RFC.  While testing this, I discovered some flaws with using the timestamp for EXDATE and RECURRENCE lookups.  As a result, the lookup key has been changed to be a pure date string, instead of the date+time ISO string.  Added tests, example code, and documentation showing how to use this.